Policyの設定と権限の機能テストを追加

+++ app/Http/Controllers/TweetController.php
@@ -22,6 +22,8 @@ class TweetController extends Controller
    */
   public function index()
   {
+    $this->authorize('viewAny', Tweet::class);
+
     $tweets = $this->tweetService->getAllTweets();
     return view('tweets.index', compact('tweets'));
   }
@@ -31,6 +33,8 @@ class TweetController extends Controller
    */
   public function create()
   {
+    $this->authorize('create', Tweet::class);
+
     return view('tweets.create');
   }
 
@@ -39,6 +43,8 @@ class TweetController extends Controller
    */
   public function store(TweetCreateRequest $request)
   {
+    $this->authorize('create', Tweet::class);
+
     $tweet = $this->tweetService->createTweet($request->only('tweet'), $request->user());
 
     return redirect()->route('tweets.index');
@@ -49,6 +55,8 @@ class TweetController extends Controller
    */
   public function show(Tweet $tweet)
   {
+    $this->authorize('view', $tweet);
+
     return view('tweets.show', compact('tweet'));
   }
 
@@ -57,6 +65,8 @@ class TweetController extends Controller
    */
   public function edit(Tweet $tweet)
   {
+    $this->authorize('update', $tweet);
+
     return view('tweets.edit', compact('tweet'));
   }
 
@@ -65,6 +75,8 @@ class TweetController extends Controller
    */
   public function update(TweetUpdateRequest $request, Tweet $tweet)
   {
+    $this->authorize('update', $tweet);
+
     $updatedTweet = $this->tweetService->updateTweet($tweet, $request->all());
 
     return redirect()->route('tweets.show', $tweet);
@@ -75,6 +87,8 @@ class TweetController extends Controller
    */
   public function destroy(Tweet $tweet)
   {
+    $this->authorize('delete', $tweet);
+
     $this->tweetService->deleteTweet($tweet);
 
     return redirect()->route('tweets.index');
